+++ Dockerfile
@@ -14,6 +14,6 @@ FROM --platform=$TARGETPLATFORM ubuntu:20.04
 
 RUN apt-get update && apt-get install -y ca-certificates && rm -rf /var/lib/apt/lists/*
 COPY --from=build /out/depot /usr/bin/depot
-# Mimics buildkitd and buildctl cli for buildx container drivers.
-RUN ln -s /usr/bin/depot /usr/bin/buildctl && ln -s /usr/bin/depot /usr/bin/buildkitd
-CMD ["/usr/bin/depot"]
+COPY --from=build /out/buildkitd /usr/bin/buildkitd
+COPY --from=build /out/buildctl /usr/bin/buildctl
+ENTRYPOINT ["/usr/bin/depot"]
